save and restore of search history
