+++ addon/globalPlugins/EnhancedFindDialog/searchHistory.py
@@ -5,7 +5,12 @@
 # See the file COPYING.txt for more details.
 
 
+from .configUtils import getDefaultConfig, strToBool
+import addonHandler
 from logHandler import log
+import pickle
+import time
+import os
 
 
 class SearchHistory:
@@ -18,7 +23,38 @@ class SearchHistory:
 		return cls._instance
 
 	def __init__(self):
+		if strToBool(getDefaultConfig("useSearchHistory")):
+			self._loadFromDisk()
+			return
 		self._terms = []
+
+	def _loadFromDisk(self):
+		# Load the pickle file from the addon directory
+		addonPath = addonHandler.getCodeAddon().path
+		filePath = os.path.join(addonPath, "search_history.pkl")
+		if os.path.exists(filePath):
+			with open(filePath, "rb") as f:
+				data = pickle.load(f)
+				if data.get("version") == "1.0":
+					self._terms = data.get("terms", [])
+				else:
+					log.error(f"Unsupported search history version: {data.get('version')}")
+					self._terms = []
+		else:
+			log.info("No search history file found, starting with an empty history.")
+			self._terms = []
+
+	def persist(self):
+		data = {
+			"version": "1.0",
+			"timestamp": time.time(),
+			"terms": self._terms,
+		}
+		# Save the pickle file in the addon directory
+		addonPath = addonHandler.getCodeAddon().path
+		filePath = os.path.join(addonPath, "search_history.pkl")
+		with open(filePath, "wb") as f:
+			pickle.dump(data, f)
 
 	def getMostRecent(self):
 		return self._terms[0] if self._terms else None
